Add files via upload
Part of screen aid desktop application.

Screenshot.py : A simple GUI region predefined screenshot taker.
Take screenshots of **any part of the screen with great precision and without loosing image quality. **
Using a transparent window as a Visual aid for Area Of Interest. 
Just Resize to cover Are Of Interest.